+++ components/app/demo/basic.md
@@ -1,0 +1,7 @@
+## zh-CN
+
+获取 `message`,`notification`,`modal` 静态方法。
+
+## en-US
+
+Static method for `message`,`notification`,`modal`.
