--- components/app/demo/message.md
@@ -1,7 +1,0 @@
-## zh-CN
-
-获取 `message` 静态方法.
-
-## en-US
-
-Static method for `message`.
