--- components/app/demo/modal.md
@@ -1,7 +1,0 @@
-## zh-CN
-
-获取 `modal` 静态方法.
-
-## en-US
-
-Static method for `modal`.
