--- components/app/demo/notification.md
@@ -1,7 +1,0 @@
-## zh-CN
-
-获取 `notification` 静态方法.
-
-## en-US
-
-Static method for `notification`.
