+++ components/app/index.en-US.md
@@ -16,9 +16,7 @@ Static function in React 18 concurrent mode will not well support. In v5, we rec
 ## Examples
 
 <!-- prettier-ignore -->
-<code src="./demo/message.tsx">message</code>
-<code src="./demo/notification.tsx">notification</code>
-<code src="./demo/modal.tsx">modal</code>
+<code src="./demo/basic.tsx">basic</code>
 
 ## How to use
